SSM Commands (#6)
## Why is this change necessary?
Need SSM Run Commands to start and stop cloud courier. Also, a
Distributor Package to install it. Note: the AtStartup trigger was
initially used for the Task set up by the distributor package, but
empirically it didn't work during computer restarts.


 ## How does this change address the issue?
Creates the start and stop commands. Uses an AtLogon trigger for a Task
set up by the Distributor Package and checking to confirm only a single
instance of cloud-courier is running


 ## What side effects does this change have?
None


 ## How is this change tested?
On a test PC


 ## Other
Moved some files around to simplify the files exposed in the main folder
to entry-level users

Added a question to be implemented in a future PR about automatically
obtaining the executable from the github release