Create a tag on the Managed Instance holding the version of the upload agent (#7)
## Why is this change necessary?
To make it easier to install/update the agent across multiple instances
simultaneously, there's not a tag on the instances indicating the
current version of the cloud courier upload agent that's installed.


 ## How does this change address the issue?
The tag is initialized to "N/A", and then the agent updates it when it
boots up.


 ## What side effects does this change have?
The IAM Role for the instance is expanded to be able to describe
instance information and update the tag on the instance.


 ## How is this change tested?
On my local computer with the new version of the agent.


 ## Other
Fixed a typo of Start->Stop. Removed the version from the AWS resource
name of the Distributor Package

Only displays the activation code in the stack output if that instance
hasn't yet been activated; this stops a huge buildup of outputs for many
lab PCs.

---------

Co-authored-by: Eli Fine <ejfine@gmail.com>